+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_enclave_functions.go
@@ -1,11 +1,53 @@
 package docker
 
-// TODO CreateEnclave
+import (
+	"context"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+)
 
-// TODO GetEnclaves
+func (backend *DockerKurtosisBackend) CreateEnclave(
+	ctx context.Context,
+	enclaveId string,
+) (
+	*enclave.Enclave,
+	error,
+) {
+	panic("Implement me")
+}
 
-// TODO StopEnclaves
-
-// TODO DestroyEnclaves
+// Gets enclaves matching the given filters
+func (backend *DockerKurtosisBackend) GetEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+) (
+	map[string]*enclave.Enclave,
+	error,
+) {
+	panic("Implement me")
+}
 
 // TODO MAYYYYYYYBE DumpEnclaves?
+
+// Stops enclaves matching the given filters
+func (backend *DockerKurtosisBackend) StopEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+) (
+	successfulEnclaveIds map[string]bool,
+	erroredEnclaveIds map[string]error,
+	resultErr error,
+) {
+	panic("Implement me")
+}
+
+// Destroys enclaves matching the given filters
+func (backend *DockerKurtosisBackend) DestroyEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+) (
+	successfulEnclaveIds map[string]bool,
+	erroredEnclaveIds map[string]error,
+	resultErr error,
+) {
+	panic("Implement me")
+}