+++ container-engine-lib/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Created a generic `ContainerStatus` object in the `KurtosisBackend` API for representing the state of containers
+* Added enclave CRUD commands to `KurtosisBackend`
 
 ### Breaking Changes
 * The `Engine` object's status will now be a `ContainerStatus`, rather than `EngineStatus`